fix: Correct interscale ratio formula and add configurable threshold
- Fix interscale ratio formula from (wtms_s1 - wtms_s) / |wtms_s| to wtms_s1 / wtms_s
- Change defect detection criteria from R < 0 to R >= threshold
- Add --interscale_threshold parameter (default: 2.0) for user configuration
- Update related comments and documentation in wtms.py, pipeline.py, cli.py, README.md

+++ glcm.py
@@ -162,8 +162,8 @@ def _compute_glcm_matrix(
     """
     h, w = window.shape
     
-    # Quantize to reduce gray levels
-    quantized = (window * (levels - 1) / 255.0).astype(np.int32)
+    # Quantize to reduce gray levels (fix overflow issue)
+    quantized = (window.astype(np.float32) * (levels - 1) / 255.0).astype(np.int32)
     quantized = np.clip(quantized, 0, levels - 1)
     
     # Initialize GLCM matrix
